requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.24.3
tensorflow==2.15.0
scikit-learn==1.3.2
Pillow==10.1.0
huggingface-hub==0.20.3
langchain==0.1.0
langchain-huggingface==0.0.9
faiss-cpu==1.7.4
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
transformers==4.36.2
sentence-transformers==2.2.2